+++ lib/decompose.d.ts
@@ -1,6 +1,6 @@
-import type { TrueObject } from '@bemedev/types';
+import type { types } from '@bemedev/types';
 import type { Decompose } from './types.types';
-type Decompose_F = <T extends TrueObject>(val: T) => Decompose<T>;
+type Decompose_F = <T extends types.TrueObject>(val: T) => Decompose<T>;
 type _Decompose_F = (val: any) => any;
 export type Decomposer = Decompose_F & {
   strict: Decompose_F;