+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 <br/>
 
+### [0.9.0] - 2024-07-07 21:32
+
+- Upgrade dependencies : @bemedev/types
+- Fix types errors
+- test(coverage): 100%
+
+<br/>
+
 ### [0.8.1] - 2024-03-15 19:30
 
 - Fix decompose array